derived from TObject and added ClassDef. Can now do ecm->Write("name").
There is still a limitation with rootcling understanding collections of
unique_ptr's. As unique_ptr's have the same memory layout as raw pointers,
we fool rootcling and tell it ECM has collections of raw pointers. This is
only used for I/O and should work. Waiting for rootcling support of unique_ptr
collections.