Add biology modules
Biology modules are used to define behavior of a simulation object.
Each simulation object has a vector of these modules.
Besides the behavior it also defines for which events it gets copied --
e.g. if a cell divides, is the module from the mother cell copied to the daughter?

Uses Variants to get rid of heap allocations. Variants cam be used to
store unrelated types in the same container -- type safe unions

Uses mpark's implementation which is compatible to the c++17 standard

Design exploration for different implementations see:
https://github.com/BioDynaMo/biodynamo/tree/design-exploration-biology-module