Prototype of user changes propagation in Paraview
Diameter values for the cells are now updated each time a user changes
the values in the live simulation (Paraview). For that purpose
"PropDiameter" array is created in the ProgrammableFilter. This array
is then fetched from the Python script and propagated back (as a whole)
to the bdmAdapter, using the vtk's DataDescription::UserData API. The
cell values are then updated in the simulator itself.