Early prototype of distribution runtime
Based on ZeroMQ library and high level C++ bindings provided by zmqpp.
The Majordomo pattern is used for basic reliability between
workers and the broker/queue, with some modifications. Timeout
heartbeats are used from broker to workers to monitor their health.

This prototype can handle worker failure (crash and restart), over
a defined time interval. When the worker remains disconnected for a
long duration, the broker removes him from the list of active workers.
Note that if the worker crashes and reconnects before the
expiration time, it will receive the undelivered pending messages.