Replace `Worker` with `BrokerCommnicator` class
This class is intended to be used as an abstraction over the connection
to the broker, by another higher-level class. Thus it is responsible
only for handling the connection to the broker and sending/receiving
messages to/from it. It also makes use of a shared (accross distributed
runtime classes) state, to perform some of its operations, e.g. sending
a message to the application directly.