Early prototype of high-level distributed API
This high-level API is used directly by the worker(s) that are
responsible for the actual computations. The API abstracts most of the
info regarding the distributed environment. This enables the (almost)
transparent worker-broker & worker-broker communication.

Internally it uses a reactor that polls all the communicator classes
(each one responsible for communication to a single node). Currently
the application receives the messages by polling a ZMQ PAIR socket.
In the future the application might call methods of this class to get
the necessary messages (abstracting all the communication).