Worker to worker communicator
Implementation of worker to worker communicator class that handles the
connection between neighbouring (co)workers. As in the case of broker
communicator class, this class abstracts the details of the connection.

Currently if a worker goes offline for some reason, the other worker
cannot detect it. Support for this kind of reliability will be
added in the near future.