WIP: Fix grid segmentation violation
The cell_division_module would crash after a certain amount of
simulation steps. This is because the grid dimensions were
defined as double precision floating points. When cells would be
at the border of a box, they would incorrectly be assigned to an
incorrect box (worst case: padding box), potentially causing a
segfault.

This is resolved by defining the dimension as integers. This
prevents the segfault and increases accuracy in box assignment of
cells.