Zero-copy message sending/receiving
No intermediate copies of the sending/receiving messages are created
during the execution of the distributed runtime. To achive this, we
switch from the ZeroMQ PAIR socket communication (used to exchange
data between threads) with mutex-protected queues.

One more reason for this change, is that we no longer have to
serialize the message, to be sent over the pipe. We serialize a single
time -- when sending the message over the network.