Client message propagation done in communicator
Instead of passing the sender (client_id) of request, that passes
through the broker, to the application we keep all senders in a queue.
This way, every time we reply to a request, we know where it came from.

Only limitation to this approach is that we need to reply to the
messages by the order they have arrived, which is logical to do
anyways (FIFO order).