Simple client management interface
Clients can now send requests to broker, in order to be perfom some
actions. Currently only two actions are supported: `CheckWorker`
(checks if a specific worker is alive or not) and `ReqTermination`
(signals the broker to terminate by the time no workers are connected)

Those actions are actually RPC calls to the broker, which replies with
ACK or NAK, whether the action was successful or not. The application
is supposed check the reply, and act accordingly.