Message contain application header
Application header can be used to create high-level logic for node
coordination, abstracting the low-level communication details. Example
of use can be the requesting and sending the "halo regions" as well as
implementing the recovery algorithm (using maybe some kind of
distributed consensus algorithm).

In addition, `SendDebugMessage`, `ReceiveDebugMessage` &
`ReceiveDebugMessageFromAny` are implemented to be used for testing.
These methods using internally the `WaitForMessage(FromAny)` methods
which return the requested type of message, if that is available.