Implement extracellular diffusion as cartesian grid
Add support for allowing simulation objects to secrete extracellular
substances. For each substance a cartesian grid is created, in which
diffusion of the concentration takes place. At any given point in
the simulation it is possible to retrieve the concentration and
gradient of a substance at a specified location through
`DiffusionGrid::GetConcentration(position)` and
`DiffusionGrid::GetGradient(position)`, respectively.

A demonstration program `diffusion_module` is added that makes use
of the diffusion grid.