Replace runtime profiling output with aggregative statistics
Instead of showing the profiling information (e.g. execution of
each operation) at every timestep, we aggregate this data, and show
a statistical overview after the simulation. The overview for now
is the total execution time of each operation.

Future idea would be to use ROOT to extract more detailed insights
from the data.