Simplified substance initialization a lot
Instead of letting the user work with the diffusion grid voxels,
we let them work with the real coordinates. This makes creating
new substance initializers a lot easier.

Also created a separate integration test for the substance
initialization, showcasing the predefined models.

Also fixed the undesired growth of the diffusion grid when the
dimensions are bound. Now the bound should always be applied if
set.