Initialize extracellular substance concentrations through user-defined functions (#26)
* Instead of choosing along which axis to initialize a substance in
the diffusion grid, we do this in the initializers, because not all
initializers need to have this axis-oriented property

* Set diffusion grid parameters in substance initializers

* Simplified substance initialization a lot

* Instead of letting the user work with the diffusion grid voxels,
we let them work with the real coordinates. This makes creating
new substance initializers a lot easier.

* Also created a separate integration test for the substance
initialization, showcasing the predefined models.

* Fix the undesired growth of the diffusion grid when the
dimensions are bound. Now the bound should always be applied if
set.

* Create config parameters for diffusion grid boundary conditions

* Add documentation for substance intilization in user guide

* Fix out of bounds bug

* When imposing a bounded box around the simulation space, cells could
be set to the right bound value, which would lead to indexing a
non-existing diffusion box. This is avoided by creating a small
distance from the right bound value when applying the bounding box.

* Apply C++ style guide

* Add unit test for substance initialization