Fix undefined symbols to OpenCL on systems without OCL
Some symbols would make it to libbiodynamo.so, which caused
undefined reference errors to these symbols on systems that do not
have OpenCL installed (during linking phase). These symbols were due
to libbiodynamo.so depending on for example biodynamo.cc, which
included header files of OpenCL (containing those symbols).

The solution in this commit is to not generate any OpenCL code in
the header files that are used for libbiodynamo.so by not defining
the USE_OPENCL variable in CMake, and only setting this variable for
the binaries (i.e. tests and demos), and of course-of-source-built
binaries.