Add Euler method in diffusion grid
The Euler method is a more accurate way of modeling
extracellular diffusion. Added convergence tests to verify
the results with the analytical solution, for varying resolutions.

In contrast to the diffusion methods used before, the resolution
of the diffusion grid results in the expected behavior. The
performance has not been affected, Therefore the Euler method is
set as the default method for solving the diffusion equation.