Refactor whole installation procedure
Remove snap and brew installation and the difference to the "development installation".
One installation method for Linux and OSX, user and developer.

The key idea of snap to build one package that works across distributions due to
encapsulation of all required dependencies (sort of containerization) works well
to ship whole applications. However, we have to ship a whole development environment.
This leads to the following issues:

  * How to add additional libraries from the host system without conflicts
    (if the core snap package is not identical to the host)?
  * What about drivers?
  * Inconsistency with development installation
  * Tedious development if libraries use hardcoded paths that point outside the snap

Therefore, we decided to rebuild the installation procedure.
Design goals:

  * One way to install on Linux and OSX
  * After the installation the usage should be the same for "users" and "developers".
    Avoid the following in the documentation: "If you are on Linux and have installed
    the user installation: execute X; if you installed the developers version; execute Y..."
  * Make it easy to support multiple linux distributions and OSX
  * Reuse code to improve maintainability.
  * Add test scripts to validate correct installation on all supported systems

https://trello.com/c/mLlZBkaC
https://trello.com/c/AYH7k25X