Leave an extra whitespace in linkflags.
The code that obtains compile and link arguments is very fragile.
There was a problem with obtaining linkflags in Travis 18.04. The
result was

  -rdynamic-L/home/testuser/.bdm/...

Notice that there was no whitespace between -rdynamic and -L...
This commit leaves a whitespace in the linkflags pattern to be
stripped off so that the -L is separated from the rest of the
command line.