Update Ray scheduling to work on solo mode.
In handle_task_waiting_cost, there's a condition to skip a local
scheduler if it is where the task comes from. In a solo cluster,
this means that no scheduler will be available to take that task.
This patch fixes that condition by using
handle_task_waiting_random when there is only 1 local scheduler.