Only save required surfaces.
Instead of saving all 27 surfaces from a box, this commit only
saves the required surfaces. For example, in a 2-1-1 scheme,
there are only two boxes on the x-axis. The left box should only
save itself, and its right surface, and the right box should
itself and its left surface. That would cut 27 writes to Plasma
down to 2.