Add the concept of events
If a new simulation object (e.g. Cell) is created during
a simulation we denote it as an event.
An example is cell division.

Simulation objects and biology modules have to define
a constructor and event handler to tell BioDynaMo how
its data members should be initialized.

This is an important change to support extensibility.