Update expected values for displacement tests
Previously, the displacement op was caching the results and applying
them after all updates have been calculated. Thus, all sim objects
observed the same state of their neighbors.

Now all operations update the sim_objects in-place and the execution
context is responsible for the discretization.

Validated that the changes are only due to in-place updates and no
other effects!