Fix visualization issue if not all objects are available on iteration 0
ParaView was crashing if a simulation object type or diffusion grid
did not exist at iteration 0. This commit fixes this problem.
A warning is emitted if if the user selects a simulation object type
or diffusion grid for visualization, but it never exists during the
simulation.

Furthermore, this commit reactivates the catalyst tests.