Add GUI visualization
The `simulate model` button will now generate code and try visualizing
the model within the GUI. It does this by generating a backup file when
initially running the model (with 1 simulation step), then restores it
into the context of the GUI for actual visualization. Current default
behaviour for user-created cells is for all cells to diffuse towards
the first cell, namely `Cell_1`. This change also includes various fixes
for issues regarding loading models and cell values.

One critical issue that needs to be sorted is that the initial backup
does not work properly. The necessary backup file is not generated
even with the correct arguments. This will hopefully be fixed in the
next commit.