Remove attempt to override preinstalled ROOT in CMake. The user needs to source the right ROOT version prior to installation if there is already another version of ROOT on the system.
This is necessary because the first call to find_package(ROOT)
sets some variables that the second call (after we download our own
version) does not seem to override (e.g. ROOT_LIBRARIES). I believe
this is because the second find_package(ROOT) is done in "Config Mode"
and behaves differently than the initial (Module Mode) call.

Either way, setting some of the environmentals manually (that should
be done by thisroot.sh is not a good practice and requires extra
maintainance to stay in sync with thisroot.sh in the future.