Add launcher script to bypass DYLD_LIBRARY_PATH problem on OS X
On OS X the DYLD_LIBRARY_PATH is stripped away from new proccesses
and therefore the @rpath of our installation could not be found.
This was the case for instance with our CLI (that opens a new
subprocess in Python to launch a simulation). On OS X we work
around this issue by sourcing thisbdm.sh prior to launching
a simulation.