Additional CMake Improvements (#76)
* Rename ROOT_DIR to BDM_ROOT_DIR.

* Check also for ROOTSYS when using find_package(ROOT);
* If both ROOTSYS and BDM_ROOT_DIR were to be found then only BDM_ROOT_DIR
  will be used. A warning will be issues to the user if this happens.

* Fix warning messages coming from CMake.

These warnings were issued only on MacOS.

* Raise a fatal error and a warning if the download of external libraries fails.

* Check if dot is installed in the system and issue a warning if it is not.

* Add graphviz package to MacOS prerequisites.

* Update documentation for NUMA and TBB libraries.

* Getting the software version for version.h now works also if the build directory is not inside the source directory.

This is archieved by passing --git-dir to the git command, e.g.
git --git-dir=/home/user/Github/biodynamo/.git describe --tags

* Add distclean command to cmake.

This way we can clear the build directory from unwanted files.

* Remove install/ directory from the build.

Everything will be done inside the 'build' directory used.

* The default install directory for biodynamo was changed from .bdm to biodynamo.

* Remove the RPath from the files inside the build directory.

This way it will be more easy to relocate the build directory and
use biodynamo from there. Not having the rpath will force us to
source anytime the thisbdm.sh script, which is however a mandatory
step in order to use the library.

* FindROOT was missing to detect fundametal library.

* On MacOS use @rpath.

This option should be enabled by default on more recent CMake versions.
We manually set it so to be sure.

* Add new method to download external packages (ROOT, ParaView and Qt5).

This method will try to donwload the packages. If the download does not
work then it will retry another time before actually aborting the
configuration step.

* Minor fixes to cmake procedure.

* Fix FindTBB warning;
* Remove commented code inside travis-osx prerequisites;
* Better detection of ROOT C++ version (using ROOT_cxx14_FOUND).

* Add extra check for `cmake3` alias

Some systems (e.g. CentOS servers) have an alias for CMake installations
with version 3 (i.e. `cmake3`). We will check if that's the case, and if
it is, we use the alias during installation.

* Remove attempt to override preinstalled ROOT in CMake.
The user needs to source the right ROOT version prior to installation
if there is already another version of ROOT on the system.

This is necessary because the first call to find_package(ROOT)
sets some variables that the second call (after we download our own
version) does not seem to override (e.g. ROOT_LIBRARIES). I believe
this is because the second find_package(ROOT) is done in "Config Mode"
and behaves differently than the initial (Module Mode) call.

Either way, setting some of the environmentals manually (that should
be done by thisroot.sh is not a good practice and requires extra
maintainance to stay in sync with thisroot.sh in the future.

* Remove search for Qt prior to search for ParaView.
We are downloading Qt anyways if ParaView is not found, so we
should first check for that.

* Remove check for CMake version and alias.
Needs rethinking on how to propagate the different binary path
to our CLI.

* Fix linking directories that lead to libraries.

* Change default installation directory to /opt instead of $HOME
Usually users clone the repository in $HOME, which leads to a
directory $HOME/biodynamo. If one would install biodynamo to $HOME
this would end up in the same directory as the repository clone.

Setting the default to /opt requires sudo permissions, however
one could also do a `make distclean` after a successfull build
and then the build directory is essentially the installation path.

* Fix `make install` directory.

* Set RPATH for libbiodynamo.dylib on Mac OS in build dir

* Set ROOTSYS in case thisbdm.sh is not sourced.

* Install to ~/biodynamo-vX.Y.Z

* Explain what the regex does

* Use git command for extracting version number.

* Remove use of launcher for out-of-source builds.

* Add quotation marks for first argument.

* Fix linking directories issue for out-of-source builds.

* Update FindTBB.cmake

* Revert "Update FindTBB.cmake"

This reverts commit 84ca34507ebabe27c87869723b14373131d59649.

* Throw error if TBB is not found.

TBB is a required package.

* capture the full cmake command line so the original command can be easily re-executed.

This can be done via the recmake_*.sh scripts that are created.

* fix typo in path of CaptureCommandLine.cmake script.

* Fix checks if cmake option are consistent between sim and bdm

* Show download progress

* Fix grammar

* Add launcher script to bypass DYLD_LIBRARY_PATH problem on OS X

On OS X the DYLD_LIBRARY_PATH is stripped away from new proccesses
and therefore the @rpath of our installation could not be found.
This was the case for instance with our CLI (that opens a new
subprocess in Python to launch a simulation). On OS X we work
around this issue by sourcing thisbdm.sh prior to launching
a simulation.

* Fix "permission denied" error in CLI

* Fix hanging Travis tests due to install directory name

Also fix the system test failing on travis due to not sourcing
BDM environments

* Also update install dir in installation tests

* Update success message with right install directory

* Fix typo and add comment

* make sure that the generated recmake_*.sh scripts are executable.

* use built-in omp.h header in case the C++ compiler does not find one.

If clang++ --print-file-name=include/omp.h does not return the location of
the include/omp.h then use a local copy of this file (it's only needed for
genreflex dictionary generation).
file (happens with Fink version of clang-5)

* local version of omp.h.

* move include of CaptureCommandLine with other incude files.