Fix rootlogon's content. Add ~/.rootrc for convenience.
The include statements in rootlogon.C were not necessary. When
dictionaries are compiled, the header information is the payload
code, so rootcling knows about the classes upon loading
libbiodynamo. Also we set USE_DICT depending on if libbiodynamo
was built with it or not. This should fix the segmentation faults
for some demos in the notebooks.