Fix multithreading bug in DisplacementOpCpu::operator()(SimObject*)
This operator containes code to initialize its state when it is
invoked the first time.
This initialization was not working if multiple threads were used.
Therefore, delta_time_ was 0 in some cases -> all mechanical forces
became 0.

This fix initializes for each thread.
A better solution would be to have a function in each operation
that could be used to initialize the state for each iteration.