Generate dictionaries for template classes through genreflex.
We were doing this with BDM_TEMPLATE_CLASS_DEF.
BDM_TEMPLATE_CLASS_DEF was originally there to avoid linking
errors when the linker expects the dictionaries of template
classes. However, it turns out that this break the one-definition
rule. One library could contain a class definition of the template
instance WITH the dictionary entries, while another library could
not (due to the template instance not being found in the header
files). It is better practice (for now) to give the initial linking
error to inform the user that there is no dictionary entry defined
for a template instance that he might use in his code.