Improving Runge-Kutta method documentation
Added extra documentation for clarity on how the Runge-Kutta method operates
in addition to the duty that the diffusion_step parameter performs within
the method.

Added additional documentation for how one can initalize a DiffusionGrid() with
specified diffusion_step parameter.

Clarified for users that the current implementation of the chemical diffusion equation
does not improve in accuracy with an increasing number of Runge-kutta steps.
Also clarified to the user that for any ordinary differential equations they use where
this method can benefit them that there is a trade off between computational speed and
accuracy.