Fix cmake RPATH settings for macOS and introduce -Drpath option for Linux.
On macOS the rpath is now correctly to be relative to the executable.
The whole tree (bin, lib, etc) is relocatable and there is no need to
set DYLD_LIBRARY_PATH anymore.
With the new option -Drpath=ON also Linux executables will get the
rpath set, however the code is not relocatable as the rpath is absolute,
however there is no need to set LD_LIBRARY_PATH anymore.