Add support for flaky tests
Flaky tests will be executed up to three times.
If one of these runs passes, the test is marked as success.

Indicate a test case as flaky by prefixing it with FLAKY_
e.g. TEST(FLAKY_Simulation, Constructor)

This features is necessary because on github actions,
tests that use xvfb fail occasionally.