Add extra checks to make sure the installed ROOT, ParaView and Qt are compatible.
To achieve this the sha256 for each third_party package is stored in the
respective third_party directory. During cmake the current expected sha256 is
compared to the stored sha256 and when not the same the third_party package
will be re-downloaded. This should reduce the number of installation errors.

On macOS we make sure not to search for Fink and MacPorts packages. Only
Brew packages are considered. This should reduce installation errors
on systems with multiple conflicting package managers installed.

The cleanbuild target did not work in case there was no third-party directory.