Improve copy command for gtest files
On some systems copying $GTEST_SRC/include would not copy the
contents of the include directory, but rather the include
directory itself - resulting in: biodynamo/build/include/include

This would in return result in Cling not being able to find the
gtest headers and cause a segfault.

The improved copy command will always copy the contents correctly.