Add googletest support for simulations (#187)
Adds support for google tests in the simulation such that users can easily write (unit) tests. 
With this low entry barrier, we motivate them to use a test-driven development approach. 
Tests are very crucial especially when dealing with complex simulations where it is very hard 
to understand where problems arise. It will help them to obtain more reliable results with BDM.

The changes in this commit are the following:

* The build system now moves libgtest.a to the build/lib folder such that people using BioDynaMo 
   in their project can access the libraries.
* Extended the CLI with a biodynamo test command
* Added a function bdm_add_test for test compilation to UseBioDynaMo.cmake.in
*  Modified the simulation template by extending the README, updating it's build instructions 
   to account for googletests, and added some dummy tests as example.
*  Upgraded to gtest-1.11.0.
*  Added tests for biodynamo test command and the test in the template to ubunut/centos system tests.
*  Modified biodynamo new command: sim name must not be test.


Co-authored-by: Lukas Johannes Breitwieser <lukas.johannes.breitwieser@cern.ch>